refactor: shorten filenames
