+++ src/momento/responses/__init__.py
@@ -8,56 +8,50 @@ from .control.signing_key.list import (
     SigningKey,
 )
 from .control.signing_key.revoke import RevokeSigningKey, RevokeSigningKeyResponse
-from .data.dictionary.dictionary_fetch import (
-    CacheDictionaryFetch,
-    CacheDictionaryFetchResponse,
-)
-from .data.dictionary.dictionary_get_field import (
+from .data.dictionary.fetch import CacheDictionaryFetch, CacheDictionaryFetchResponse
+from .data.dictionary.get_field import (
     CacheDictionaryGetField,
     CacheDictionaryGetFieldResponse,
 )
-from .data.dictionary.dictionary_get_fields import (
+from .data.dictionary.get_fields import (
     CacheDictionaryGetFields,
     CacheDictionaryGetFieldsResponse,
 )
-from .data.dictionary.dictionary_increment import (
+from .data.dictionary.increment import (
     CacheDictionaryIncrement,
     CacheDictionaryIncrementResponse,
 )
-from .data.dictionary.dictionary_remove_field import (
+from .data.dictionary.remove_field import (
     CacheDictionaryRemoveField,
     CacheDictionaryRemoveFieldResponse,
 )
-from .data.dictionary.dictionary_remove_fields import (
+from .data.dictionary.remove_fields import (
     CacheDictionaryRemoveFields,
     CacheDictionaryRemoveFieldsResponse,
 )
-from .data.dictionary.dictionary_set_field import (
+from .data.dictionary.set_field import (
     CacheDictionarySetField,
     CacheDictionarySetFieldResponse,
 )
-from .data.dictionary.dictionary_set_fields import (
+from .data.dictionary.set_fields import (
     CacheDictionarySetFields,
     CacheDictionarySetFieldsResponse,
 )
-from .data.list.list_concatenate_back import (
+from .data.list.concatenate_back import (
     CacheListConcatenateBack,
     CacheListConcatenateBackResponse,
 )
-from .data.list.list_concatenate_front import (
+from .data.list.concatenate_front import (
     CacheListConcatenateFront,
     CacheListConcatenateFrontResponse,
 )
-from .data.list.list_fetch import CacheListFetch, CacheListFetchResponse
-from .data.list.list_length import CacheListLength, CacheListLengthResponse
-from .data.list.list_pop_back import CacheListPopBack, CacheListPopBackResponse
-from .data.list.list_pop_front import CacheListPopFront, CacheListPopFrontResponse
-from .data.list.list_push_back import CacheListPushBack, CacheListPushBackResponse
-from .data.list.list_push_front import CacheListPushFront, CacheListPushFrontResponse
-from .data.list.list_remove_value import (
-    CacheListRemoveValue,
-    CacheListRemoveValueResponse,
-)
+from .data.list.fetch import CacheListFetch, CacheListFetchResponse
+from .data.list.length import CacheListLength, CacheListLengthResponse
+from .data.list.pop_back import CacheListPopBack, CacheListPopBackResponse
+from .data.list.pop_front import CacheListPopFront, CacheListPopFrontResponse
+from .data.list.push_back import CacheListPushBack, CacheListPushBackResponse
+from .data.list.push_front import CacheListPushFront, CacheListPushFrontResponse
+from .data.list.remove_value import CacheListRemoveValue, CacheListRemoveValueResponse
 from .data.scalar.delete import CacheDelete, CacheDeleteResponse
 from .data.scalar.get import CacheGet, CacheGetResponse
 from .data.scalar.increment import CacheIncrement, CacheIncrementResponse
@@ -66,14 +60,14 @@ from .data.scalar.set_if_not_exists import (
     CacheSetIfNotExists,
     CacheSetIfNotExistsResponse,
 )
-from .data.set.set_add_element import CacheSetAddElement, CacheSetAddElementResponse
-from .data.set.set_add_elements import CacheSetAddElements, CacheSetAddElementsResponse
-from .data.set.set_fetch import CacheSetFetch, CacheSetFetchResponse
-from .data.set.set_remove_element import (
+from .data.set.add_element import CacheSetAddElement, CacheSetAddElementResponse
+from .data.set.add_elements import CacheSetAddElements, CacheSetAddElementsResponse
+from .data.set.fetch import CacheSetFetch, CacheSetFetchResponse
+from .data.set.remove_element import (
     CacheSetRemoveElement,
     CacheSetRemoveElementResponse,
 )
-from .data.set.set_remove_elements import (
+from .data.set.remove_elements import (
     CacheSetRemoveElements,
     CacheSetRemoveElementsResponse,
 )
