+++ src/momento/responses/data/list/pop_back.py
