+++ src/momento/responses/data/list/pop_front.py
