+++ src/momento/responses/data/list/remove_value.py
