+++ src/momento/responses/data/set/remove_element.py
